backend/requirements.txt
boto3==1.20.11
fastapi==0.70.0
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.15.0